+++ lib/mongoid/criteria.rb
@@ -23,6 +23,12 @@ module Mongoid
     include Modifiable
     include Scopable
 
+    # Static array used to check with method missing - we only need to ever
+    # instantiate once.
+    #
+    # @since 4.0.0
+    CHECK = []
+
     attr_accessor :embedded, :klass
 
     # Returns true if the supplied +Enumerable+ or +Criteria+ is equal to the results
@@ -322,7 +328,7 @@ module Mongoid
     #
     # @return [ true, false ] If the criteria responds to the method.
     def respond_to?(name, include_private = false)
-      super || klass.respond_to?(name) || entries.respond_to?(name, include_private)
+      super || klass.respond_to?(name) || CHECK.respond_to?(name, include_private)
     end
 
     alias :to_ary :to_a
@@ -506,8 +512,10 @@ module Mongoid
         klass.send(:with_scope, self) do
           klass.send(name, *args, &block)
         end
-      else
+      elsif CHECK.respond_to?(name)
         return entries.send(name, *args, &block)
+      else
+        super
       end
     end